Personal Projects/PDF to csv/requirements.txt
tabula-py==2.7.0